Added support for global passes  to the pass managers
These type of passes can be used to integrate in the pass manager framework optimization that need fine grained control on how they walk the tree (for example they act on the node childrends rather than on the node itself, and so on).

In this case the pass is only called once with the entire AST as an argument, and it implements its own tree walking strategy. The function should then return a pointer to  the processed tree.